scheduler: process front tasks in batches
There was a problem that if the scheduler takes millions of tasks
at once, then handling all of them coming from the front queue
could take 100-200 ms. During that time only one thread was
working, processing the tasks. The others were idle, because the
tasks couldn't be pushed to the ready-queue right away.

The patch makes so the scheduler processes front and waiting tasks
in batches of a limited size. Also the worker threads perform
scheduling more frequently, not just when the ready queue becomes
empty. It helps to prevent the front-queue from growing too much.

The patch greatly improves parallelization for the described case
(tens of percents speed up) and doesn't affect the cases where the
load is not so extreme.

Part of #4