bench: introduce benchmark implementations
The patch brings in the benchmarks for TaskScheduler and
MultiConsumerQueue.

The benches not only allow to measure absolute values on various
workloads, but also can compare the original classes with their
trivial versions: a mutex-based task scheduler variant, and a
mutex-based multi-producer-multi-consumer-queue.

The benches can be run locally individually, or multiple times to
get aggregated statistics. But most importantly - there is a
script which allows to run great number of benchmark scenarios and
generate a markdown report.

See the bench/README.md file for more info.

Part of #4