threads: use std mutex and condition_variable_any
These are platform-agnostic and portable. Also they seem to give a
notable perf boost compared to the raw APIs. The benches are to be
fully re-evaluated after some more patches about migration to the
standard library.

Note, that the perf increase (got ~6% on Mac) doesn't look related
to the removal of PTHREAD_MUTEX_RECURSIVE attribute. Removal of
this attribute alone also improved things a bit, but not that
much.