scratchpad: drop it
It is replaced with mg::common::StringFormat and StringVFormat.
They basically do the same, but return std::string.

This is a part of the ongoing activity to drop HybridString.
String performance doesn't matter for TaskScheduler, so it is
better to just use std::string instead of maintaining an own
string container.