array: drop it
Use std::vector. Arrays are used for insignificant things in the
scheduler. Whatever implementation is used, it doesn't affect
performance.

On the other hand, dropping the custom Array allows not to bother
with maintaining it and its tests. It will also simplify usage of
the scheduler directly as a library in the future.

Benchmarks didn't show any difference on Linux in a cloud.

On Mac though any attempt to use std::vector in Bench.cpp and
Bench.h (even without all the other changes) led to visible slow
down. Note that it happens even if the vector is not used in the
scheduler itself, but only in the bench code. Looks like a
Mac-specific problem, maybe related to how libc++ is linked here
or something. Not important.