mpsc-queue: relax some memory barriers
Same as with MultiConsumerQueue, the MultiProducerQueue didn't
need sequential consistency for either of its atomic operations.

The patch relaxes the barriers to acquire and release ones.

In a few benches the perf win is up to 15%, but mostly it is ones
of percents or zero. Barriers relax isn't too helpful when all
the contention is for the single memory spot anyway (myHead).

Part of #16