scheduler: relax some memory barriers
Neither of the atomics in the scheduler needed sequential consistency. All the
atomic usecases are trivial except the task status. This one most of the times
can be accessed in a relaxed way. The reason is that all the tasks go through
the front queue via whose acquire-release barriers their memory is synchronized
in bulks.

The only exception is the signals which can be sent and received completely
bypassing the queues. Those should use acquire-release.

There was observed up to 20% speed up for the benchmark nano load, 2 threads,
5 000 000 tasks, each is executed 10 times. The others seem either a few
percents faster or inside the noise level.

Part of #16