cmake: add installation commands
Previously the code had to be manually built and the resulting libraries +
headers had to be copied into other projects. Especially it was problematic for
the headers not all of which are needed outside of this project.

Now cmake has proper install commands which can either copy the needed files
into a global place (like /usr/lib and /usr/include) or into a local folder.